Printing images to screen
Fixed a lot of syntactical errors and I now understand Tkinter and
Pillow enough to  actually create a window and display the generated
images on it.  A lot more happened but it took me so long to figure all
of this out that I forgot many of the other things that I've done since
the last commit.

From here the next major task will be "animation" but there are many
other small things to do as well, I'm sure.