Beginning of animations
Window now has correct size

rowNum variable is now a member of room

Now supports "animation" data in csv file, which is stored directly
following the end of the description of the last object. (makes use of
itertools)

Contains lots of unneeded variables (rowLen, charNum just to name a
few).  This is because I'm currently using the csv package (package) to
read the files but if I run into any major efficiency issues down the
line I will want to switch to normal file read/write stuff and index
based on the character number, not the line number.  More of a
note-to-self at this point, but yeah.

Let's see, what else...
I need to update the readme for sure, with further instructions

Major issue:  Gets to end of file and then can't go anywhere from there

Also, I need to write functionality for going backwards in the list.

Honestly, at this point it seems like csv module is gonna be too slow
even at this scale.  But the limiter could be something else entirely,
like tkinter.  we'll see how it goes.