Started C++ work
It was a little difficult to get into this after doing nothing with C++
for almost 3 months, but I seem to be getting the hang of it pretty
quickly (at least, a lot more so than I did at the beginning of last
semester after a 7ish month break)

Anyway, I started work on the C++ file(s) for generating .csv input to
play.py

Currently one major issue is that I'm passing everything by value,
including arrays that could reach a decent size, so I'm gonna need to
deal with that later.

Actually I just got a fair bit into coding (line 201 on datagen, to be
exact) when I realized I'm pretty much gonna have to handle the data
arrays as dynamic memory.  So i'm gonna call it a night, go to bed, and
maybe do all that if I have time tomorrow.