datagen.cpp supports file input
Could be more clear if I simply overload the stream input operator.

Anyway, now supports the use of a "datagen.csv", which I definitely need
to upgrade the readme.md to explain.  But it's pretty self explanatory.
datagen.csv is used, through datagen.cpp, to generate input.csv, which
is interpreted by play.py to generate the "scene"