Simulation class
Main function was getting way too long, so I broke it down by writing a
simulation class as well as "inCSV" and "outCSV" functions.

The simulation class holds the x and y resolutions of the window, the
number of objects, and then 2 arrays: colors and objects.
colors is an array of pointers to Color and of size numObjects, and
objects is an array of pointers to Object and of size numObjects.

Program (datagen.cpp) seems to run MUCH slower than before, even though
the data generated is the same.  So I definitely need to look into that.