Physics!
Major work done on the physics calculations.  Other than tons of new
member variables and get/set functions in both Object and Simulation,
major changes include:

Calculations of:
- Mass
- Distance
- Gravitational force
- Acceleration
- Velocity (instantaneous and change)
- Position (instantaneous and change)

Datagen.csv has been updated to contain 2 objects representing the earth
and the moon, approximately to scale regarding size.  Once I implement
the carrying out of calculations and output of animation data, I'll be
comparing this with some calculations that I did by hand in order to
confirm accuracy.

At the current moment, this is built to only work (correctly) in
2-object systems.  So one next major change would be to support
calculations for systems with 3+ objects
Another change I need to make is to allow both simulation constants
(like time interval) and object constants (like density) to be defined
from the datagen.csv file

Note that this commit doesn't perform any calculations, rather it just
has the ability.  Actual implementation of that as well as expanded
output should hopefully come in the next commit.