More physics work
More work done on the actual calculations.  A lot of inefficient code
and some of the functions written in the previous commit are obsolete.
I'll make sure to go back and clean that up later, especially when
dealing with efficiency issues.

I also need to sort out the issues with whether density should be a
member of object or simulation, and then also maybe implement friend
variables so that the Object class has access to simulation's pixelSize
variable, or find some other way to deal with that.