More efficient calculation
Cut off a good amount of time from the force calculations.  To be
precise, the program is now more efficient by a factor of (1 -
t(n))/(n^2)), where t(n) is the nth triangular number and n is the
number of objects in the system.

Wrap around disabled in play.py

I'm not sure if the gravitational physics is broken, I haven't done
enough looking into it to be able to tell.  But assuming it works, then
today I fixed both the memory leaks and the inefficient force array
generation.

I also tweaked the positions in the datagen.csv file to allow easier
viewing of the 3 objects in case one were to go off screen.

I am relatively sure something is still wrong with the calculation side
of the code because while there are no memory leaks, something else is
up.  I don't think I'm gonna able to be too productive at this point in
the night but I'll certainly revisit this later, even if I don't have
time to over the next few days.