Debug mode and more specifications
More things are now specified in the datagen.csv file.

The first line is now
width, height, numObjects, pixelSize, timeInterval, numIterations,
debugMode
where pixelSize is the length or height for a pixel to represent (in
meters), timeInterval is the delta(t) to use for velocity and position
calculations (in seconds), numIterations is the integer number of frames
to generate, and debugMode is a string representing whether to run the
program in debug mode.

If debug mode is enabled, the program continuously outputs all sorts of
data on the calculations being done and inputs being received
Specifying debugMode as "true", "y", or "debug" will enable debugMode,
while "false", "n", or "nodebug" will disable it.  Case is not
important.

I'm thinking of changing debug mode to be a third argument specified
when running the program, instead of making it a part of the datagen.csv
file.  But I'm not too sure at this point.